requirements.txt
Flask==3.0.2
flask-socketio==5.3.6
opencv-python==4.9.0.80
ultralytics==8.1.28
torch==2.2.0
Pillow==10.2.0
numpy==1.26.4
python-engineio==4.9.0
python-socketio==5.11.1 